Fix for Ripple on Fast Thumbnail Indicator
There was a cute optimization was to start the ripple animation, and
then fill in the bitmap in 30ms after the bitmap was computed.
However, when the system is overloaded, the bitmap is not necessarily
computed in 30ms, and an empty ripple can result.   Put in a more
conservative fix, when the ripple only starts when the bitmap has
already been computed.

Bug: 19132963
Change-Id: Id39257a1d75342c78385bac8682462e8875d97f7