Make sure to always use ZSL frames for capture.
  Bug: 19166357

This fixes an issue on the N9 which does SW NR with
TEMPLATE_STILL_CAPTURE, which stalls the whole camera image reading
pipeline for a long time.

And in general we should probably not take pictures with potentially
differenty qualities for ZSL and while waiting for 3A to converge.

M will solve this issue with the introduction of the reprocessing
pipeline.

Change-Id: Ic444884f5dad6ecff44d0a6b7c7a118f0242bef2