Remove UNKNOWN Orientation values
UNKNOWN Orientation is causing some strange crashes in the Camera App,
in certain situation when the phone starts up while faceup on a table,
i.e. an unknown orientation.   This CL is to make sure that we stop
any UNKOWN orientations propogating through the system and causes these
sorts of crashes.

Bug: 19049293
Change-Id: I4f403bda1b6b23f2bbe5ee590e23d0fc078543b7