Make sure to null SurfaceTexture when destroyed.
  Bug: 19149830

This will prevent unnecessary camera open events. We assume that when a
SurfaceTexture is present at the time we resume, that it is usable and
we can initialize the camera. Not nulling it causes the camera to be
initialized on a potentially invalid Surface. Once the Surface becomes
available, we'd close and re-open the camera, which slows down our
resume times.

By nulling the SurfaceTexture when it gets destroyed we make sure to not
initialize the camera in resume an instead wait for the proper callback.

This might get rid of the HDR+ resume issue we've been seeing as well.

Change-Id: I71837aeda79c2f7516e64bdb10c12d39f728c0ca