Cancel implicit AE lock after AE precapture sequence
After taking, or attempting to take a flash photo, the app currently
sends a CONTROL_AF_TRIGGER_CANCEL to release the AF lock.  On some
devices (N6), there appears to be an implicit AE lock set by the
CONTROL_AE_PRECAPTURE_TRIGGER, which is unlocked when a flash photo
is captured (with INTENT_STILL_CAPTURE).  In the event that the app
is interrupted before sending the still capture request, it must
still cancel the implicit AE lock.  As a workaround, sending
a request with AE_LOCK=true followed by AE_LOCK=false appears
to reset this implicit AE lock.

Workaround for Bug: 19265647

Change-Id: Ie76a81c1a138fb2334de736b315796e869378bf4