Remove app background and reduce 1 layer of overdraw
Set the app background to null. This saves one full layer of
overdraw in the filmstrip. The capture UI is more dependent
on the window background, both for gaps in the preview and
screen edges, and behind the shutter button when it is
made transparent when the mode list is open. The preview
is a large portion of screen and to remove one overdraw,
we instead fill gaps around the preview with a
MarginDrawable overlay. This CL leaves a background behind
the shutter button when opaque, to allow for it becoming
transparent during mode list view.

Bug: 18596159
Change-Id: I2a9083a2413944c6244e2f0528164e8852ae3d8e