Route all thumbnail updates through CaptureSession
  Bug: 19326495
  Bug: 19310525

This removes the thumbnail ready callback that was hooked directly into
the ImageBackend from the CaptureModule before since it's now
unnecessary.

Client of CaptureSession can no optionally specify whether the updated
thumbnail is to be used to update the capture indicator. Default is
'yes' so that Lens Blur, Photo Sphere and HDR+ don't need to change.

Change-Id: I94adb50f875ccee214492f03f00ed2a67d400da1