N4 CaptureModule not closing
N4 CaptureModule was not wired up properly to close the device when it
closed.  There's still an issue that the synchronization is incorrect,
i.e. we don't guarentee close has completed before we open it again.
There's a fix so that we can use OnCameraAvailable listener to only open
the camera when it is available.

Bug: 19332637
Change-Id: Ie88acb3a9f75a5896606dd660c005de0dc82dc3c