N4 CaptureModule Memory Optimization
Although functionally correct, there was some processing that was
wasteful of memory in N4 Capture Module.   The two major culprits
are the implicit full-size JPEG decoding for the thumbnail, and
using the capacity (instead of the limit) of the byte buffers when
making a memory copy of the JPEG data.  Fixes include a explicit
subsample value for thumbnail and a tighter memory allocation based on
the limit.

Bug: 19354120
Change-Id: Ibd94c30974753e426254ac002e795b5d1de8bcba