N4 Capture Module Rotation/EXIF fix
Finally, have some understanding what values are supposed to be
passed to the session vs. what are supposed to be put in the EXIF
flags.  Although they are similar, they are used for slightly
different purposes.   This CL has been tested on both N5 and N4
and seems to give filmstrip behavior and forwards the correct
EXIF values.

Bug: 19378973
Change-Id: Ie8b41b6d8e03e0a5392da8f0e7b52f65383b4e4f