Add a PreviewSizeSelector interface.
This removes the required context object for the utility objects
in both CameraUtil and CaptureModuleUtil. Adding an interface
for PreviewSizeSelection clears the way for future work to move
the preview size selection code into a dedicated class and to have
api specific versions of preview selection code for startup
speed improvements.

Change-Id: I92e58dad23b08057eb10984c31d04947a376ac6d