Centralize logic for Camera API feature determination.
  Bug: 19179276

This captures camera capabilities in a centralized class and allows
them to be overridden by GServices in the field or using adb.

Since this is not it's own class it would be trivial to create another
instance for testing and pass that into CameraActivity or even to change
these values using a UI before the actual Activity is started.

Change-Id: I74262c8467f631c7b5e42b02d269d5f401d4e9e7