Legacy Autotransform Support to CaptureModule
Added the legacy support for Capture Module to use
the combination of transformations, dynamic layout, and
transform helpers that PhotoModule uses.  It seems
to replicate that same functionality as the CaptureModuleUI,
but it's good to have that option for comparison, and to be
able to turn it on and off with one line configuration.

Bug: 19271661
Change-Id: I4ad4fa6fe3547373d6fbf77a53a1a18cd26df8ec