N4 16:9 Preview Workaround
The N4-specific fix is to trick the layout UI logic to accept a
16:9 preview coming from a 4:3 surface.  When we render the 4:3
surface, we fill in the size of the preview area.  So, if
the surface dimensions are 4:3 AND the surface actually
contains 16:9 content AND the preview view is also 16:9, then
preview will actually look correct.  To maintain this inconsistency
between the surface and the content aspect ratio, we use the
legacy layout path, and trick the UI layout into laying out
a 16:9 preview, but using a matrix transform that converts 4:3
surface into a 16:9 view.  In new and improved matrix transformation
codepath, we correctly assume that the matrix and the preview
size should be consistent.

Currently, the 0.9MP camera setting on the N4 on the front and back
camera produces a squeezed preview, but also produces a squeezed
JPEG artifact and is probably another independent N4 HAL related issue.
But it's WYSIWYG, so I guess it's correct.

Bug: 19271661
Change-Id: I213a423f891a7c216568970c7f520a6134f9ad13