Automated ImageProcessorListener registration
I've added the ImageConsumer interface and implementation
in ImageBackend to support the automatic registering and
unregistering of the ImageProcessingListener.  This
new functionality avoids the issues of reference leaking,
so that the transient listener implementations are NOT
held by the image backend.  Also more general purpose
interface has been created such that any passed Runnable
will be executed when the associated set of tasks that have
been enqueued with ImageBackend::ReceiveImage have finished
processing.   This functionality solves the referencing
leaking issue of ImageProcessingListener in the current
implementation.  Also, removed all references to Main
thread executor so that ImageBackend is not a direct
suspect in jank issues.

Bug: 19517380
Change-Id: I806582f13157030ed6bc23ad70a4efd93339d9ff