Limit setting photo item placeholders when session is done.
Only add placeholders for an item visible in the filmstrip. If a session
is done for an item not visible, it is less janky for the user to see the
standard gray placeholder while loading the image from disk.

Pruned older methods of determining filmstrip visibility, none of which
worked and relied on older code assuming the camera preview was an item
in the filmstrip, including 'sticky' filmstrip item bits. Added a method
to CameraAppUI that proxies the FilmstripLayout visibility, which is
actually what app uses to show/hide the filmstrip.

Bug: 19517380
Change-Id: I5e2a61b55aa2045bfd09f495dd02cdcb9c65dc68