Swap dimensions of artificial 16:9 resolution on N5 on upgrade.
  Bug: 19693226

Since the N5 doesn't natively support a high resolution 16:9 size we need
to artificially add it and then crop the result from the high-resolution
4:3 size. In version 2.4 we unfortunately swapped the dimensions of
ResolutionUtil#NEXUS_5_LARGE_16_BY_9_SIZE, which now causes a few issues
in 2.5. If we detect this case, we will swap the dimensions here to make
sure they are the right way around going forward.

Change-Id: I579e464d177ac86d6cc88e9142c26fd9d6b67a76