Fix 16:9 start-up dialog selection.
  Bug: 19696056

We need to make sure the list from which we choose the highest
resolution size of a given aspect ratio based on a list that is filtered
to be chosen by the user. In case of the N5 the native camera list will
be augmented with a new high resolution 16:9 size which needs to be
selected if the user chooses 16:9 on start-up on the N5.

Change-Id: I03a86aa0a148b6cbabbd51a577202b79b4941d7d