Re-arrange camera initialization.
This change re-arranges several things to make it possible to know
the CameraId you are about to open before you actually open the
camera device. It also adds an explicit CameraId class for handling
both legacy and camera2 id comparisons, an active camera tracker,
and moves the camera id query mechanism into a hardware manager
class in prep for the subsequent camera release.

Bug: 19679820

Change-Id: I4c09b9a12a1ea5badcac3a6a073a56764f9a7eba