Change CaptureSession.saveAndFinish to return a future for the URI.
  Bug: 19665531

Most callers of saveAndFinish never provided the callback. In addition
it's unclear on what thread the callback is called. This cleans this up
a bit and moves the responsibility to the caller to decide on which
thread the URI should be received. Callers that don't care don't have to
do anything.

Also fixes a bug where we wouldn't even supply a URI for when there was
no placeholder.

Change-Id: I1a186b470d2cca8f34a674e9eb5c533fc01a4409