Increase the thread priority of critical camera components.
This increases the priority of the frame server and metadata
response listeners that must respond immediatly to frames and
callbacks that are returned from the camera HAL. Both threads have
a tiny amout of code and callbacks that get run but must not block
or they will stall the preview on image capture. This appears to
only have limited impact on jank in the app.

Bug: 19076469
Change-Id: I41845e094feffb1ecd3be9613489a3d70f916580