Fix view recycling in filmstrip
View recycling is pretty broken, with code put in for the old
camera preview in filmstrip being used to avoid recycling in all
cases. This looks like its been broken in a number of releases, but
less so due to a destructive removal of views during updates. I had
fixed that but it then exposed the issue that recycling wasn't occurring
elsewhere. This CL removes views from hierarchy, confirmed no more leaks
when capturing and when swiping through filmstrip.

Bug: 19970885
Change-Id: Ic63b9231bd24db0f44a99567278886eb0998d740