Send a warmup burst to the camera before attaching the preview.
This reduces startup jank by submitting a warmup burst of n
frames to the camera before the preview surface is attached to
the camera. This ensures the large internal buffers are
allocated by the HAL before the preview stream is attached. The
change will delay the preview by n frames in exchange for a
smoother preview once the preview is actually shown.

Bug: 19076469

Change-Id: I5abb8b49a64c35686c014001a45239f4eccfe34d