Check for a potentially destroyed activity before startPreview.
There is a small chance that the activity can be destroyed
between the time the camera is created and the preview is started.
This can lead to unchecked NPE's across the app because most code
assumes that there will never be a case where activity objects
are accessed after onDestroy.

Bug: 20097425
Change-Id: I2904ffba451ded2417a806c83fd12e4f62611150