Refresh metadata for every filmstrip item update.
Current filmstrip item updates:
1. call CameraFilmstripDataAdapter.updateItemAt which then calls FilmstripView.updateViewItem
2. FilmstripView.updateViewItem updates the view then calls CameraActivity.updateUiByData
3. CameraActivity.updateUiByData calls CameraFilmstripDataAdapter.updateMetadataAt
4. MetadataUpdateTask calls FilmstripView.updateViewItem again if metadata was loaded

The double updateViewItems disrupts SessionItem -> PhotoItem transitions, and for some
reason for LB causes the second updateViewItem to only request tiny thumbnails. For a very targeted
fix, we instead make step 1 do the metadata load, and then updateViewItem is only called once. The
code in step 3 still executes but is a no-op since metadata was already loaded in step one.

Tested with Camera mode, Pano/PhotoSphere, LB, redoing an LB, calling out to editor and changing a
photo, and taking a video.

Bug: 20862631
Change-Id: Ib8b112c2ef1b5a364a5569d1391ef359d1b5fb1a