WAR of the face detection failure
This patch gives a work-around of the face detection failure that
occurs when the ZSL camera is used. It forces to request a preview
frame followed by a 5-frame ZSL burst before requesting the repeated
preview and ZSL streams. It also hides the first preview frame in order
to avoid potential janks. We do this only for L, Nexus 6 and Haleakala.

Bug: 20724126

Change-Id: I40b7861e07570ed53135b2d8d182b65c5799adac