Workaround of the face detection failure on L and N5
This patch is a workaround of the face detection failure on L
and Nexus 5. For Nexus 5 on L, the current FD only works when
the preview stream is configured as the first stream before the
JPEG stream. This change lock the order by creating a capture
session with the preview stream first, and then creating the
normal capture session with preview and JPEG.

Bug: b/21039466

Change-Id: Icb7f1f758bc9d7ca428c7a4748223072d39dd57f