Fix for invalidated resolutions via upgrade
Bug: 21758681

There was a upgrade bug where an valid resolution of the
previous version of the Camera App could then be invalidated
the subsequent version of the Camera App.  If there were no
other valid resolutions of the same aspect ratio available
to the app, the current Camera App would choose (0,0) as its
default size and then write this value into the saved settings.

This fix allows the current Camera App to consider other resolutions
not of aspect ratio as a fallback resolution setting.  Also, if
an invalid resolution valid has been saved to the settings, it
also ignores it and then resaves a valid resolution value to the
settings.

Change-Id: I08f19d1ee6491619e92697897b73e4fd33a19a00