Fixes to M permissions.
Previously, we called PermissionsActivity with startActivityForResult().
However, this creates race conditions as we check for permissions
and the CameraActivity continues to operate. Now, we end CameraActivity
and launch a new instance upon successful permissions resolution.
We can also put the preload filmstrip logic back in its original
place in onCreate. The checks for permissions happen in both
onCreate and onResume.

Bug: 22442745, 22478144, 22497152
Change-Id: I82e9125a46581db44aa61d4ee94aec5a820e9df0